// frontend/src/api/models/novel.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { NovelType } from "./novelType";
import type { Visibility } from "./visibility";

/**
 * Pydantic schema for novel.
 *
 * Attributes:
 *     novel_id: UUID id of novel in db.
 *     novel_title: String title of novel.
 *     novel_description: String summary or description of novel.
 *     novel_author: String author or description of novel.
 *     novel_visibility: Visibility enum of novel.
 *     novel_type: NovelType enum of novel.
 *     language_code: String code key to language of the novel.
 *     source_work_id: UUID foreign key to source work of the novel.
 */
export interface Novel {
	languageCode: string;
	novelAuthor?: string | null;
	novelDescription?: string | null;
	novelId: string;
	novelTitle: string;
	novelType: NovelType;
	novelVisibility: Visibility;
	sourceWorkId: string;
}